Fixed birthday and creationDate parsing.
In TorcGraph, birthday and creationDate are stored as date strings (the same
strings that are found in the generated datasets for LDBC), however the SNB
interactive workload driver expects values of type Long for these two
properties. Turns out these are the time in milliseconds since the default
epoch in the GregorianCalendar class, in the GMT time zone.