Moved hardcoded entity numbers to enum.
Previously, entity number assignments (used to make entity IDs go from being
locally unique within entity classes to globally unique in the graph database)
were duplicated in the the graph loader TorcGraphLoader and the database
connector TorcDb.

This information is now centralized in a new Entity enum which defines the
number and name string for every type of entity is the LDBC SNB graph.