Added commons-cli to top of dependency list.
The LDBC driver code was including an older version of the commons-cli library
which did not include the version 1.3.1 features needed by TorcGraphLoader. This
problem was introduced when moving the LDBCSNBBulkLoader from the torc-gremlin
repository to this repository.

Fixed this by moving commons-cli library to top of dependency list, relying on
maven's policy of only including the first encountered version of a library in
the resulting package.