Added more informative progress reporting output.
Loader now outputs the total time that has elapsed and the number of
lines from the currently processed file that have been loaded.

Also, the dotRate parameter that controls the period of reporting has
been renamed to progReportPeriod, since it's actually a period, not a
rate, and progReport more accurately describes the parameter's use.