Removed final modifier from method parameters.
After some deliberation, I decided to, as a rule, not use use the final
modifier on method parameters. A method parameter being final is of no
consequence to the method caller, since Java is technically pass by
value, and so final or non-final, the method is not able to affect the
value of the variables used as arguements by the caller. Declaring a
method parameter as final seems to only be of any consequence to the
method implementer, who is then forbidden from making an assignment to
that variable. Such constraint on the method implementation embedded in
the method signature IMHO violates the separation of concerns between
interface definition and interface implementation.