Removed Neo4jCypherResult.
Since exceptions are thrown in the event of errors and not returned as a
part of the execution results for a call to exec or execAndCommit, the
Neo4jCypherResult object is no longer needed to encapsulate this
information. Instead, results are reported as a 2d array of Strings,
where the 0th row stores the column headers, and rows 1 to N-1 store the
column values. Returning the results as a 2d String array (as opposed to
a JsonArray) frees the user from having to use a particular Json
library, and also from having to understand the inner JSON structure of
the returned results from the server. These details are safely hidden
from the user.