Added coalescing of person properties into person.
LDBC SNB Data Generator outputs separate files for two properties on
persons: email, and speaks. The Neo4j Import Tool, however, expects that
all properties for nodes (and relationships) are present on a single
line in the file listing all the nodes (or relationships) of a
particular type.

Now the conversion utility will do this automatically for us.