Changed return data types for TxDriver.
Neo4jTransactionDriver previously returned List<String[][]> for calls to
exec and execAndCommit, but this was problematic because then the user
is forced to figure out which column is which when they likely already
know the column titles from the original Cypher statement.

Changed the return data type to List<Map<String, String[]>> so when the
user knows the column title they can find the column directly.