Changed post and comment labels in Neo4j.
Turns out sometimes it is important to refer specifically to nodes that
are of one type or the other, and sometimes you have to refer to a
"message". To accomodate the latter with separate labels we can use a
WHERE (m:Post OR m:Comment) in Cypher queries, but accomodating the
former without separate labels is not possible.