Reworked file parsing for updates.
Previously update stream file parsing for finding update parameters was
ad-hocly parsing lines from the file for parameters to construct
instances of the corresponding operation classes. Continuing like this
would have resulted in a lot of code duplication and logic spread
throughput the code for the QueryTester.

Instead of doing that, I centralized the parsing logic into a method,
parseUpdate, that takes a line from the update stream file, figures out
what kind of operation it is, parses its parameters, and constructs an
update operation object of the right kind and returns it. Calling code
will need to know what to expect in return and that's not good, but it
does save the caller from needing to know how to instantiate an
operation of that type, and thus more completely encapsulates the
knowledge of this in the parser.