Refactored short reads 1-7.
Now using the Java reflection API to more compactly and generically
construct operation and operation handler objects based on the type
of query the user wishes to execute. The code is still not
implementation independent, but is moving in that direction. Not a
needed feature at this exact moment, but will be soon.