Changed the return types for Neo4jTransactionDriver.
Changed this to return a list of a type that exposes the Jason format of
the result, but hides the details of how that type is encapsulated in
the server's response. This type is called Neo4jCypherResult and allows
the user to fetch the data as JSON arrays, one array for the column
headwers, and an array for each row of the results.