Added support for query10.
Previously Query10 was not able to run because it depends on properties
"birthday_day" and "birthday_month" on Person nodes in order to run, but
those properties had not been included when generating the dataset files
for Neo4j.

To support this, I modified the DataFormatConverter to include these
properties in the files generated for the Neo4j import tool.

Also, Query10 from the spec had a slight bug in its calculation of the
number of the next month. If month numbers go from 1 to 12 (as I assume
is the case because day numbers appear to be numbered starting from 1),
then the next month is calculated: (m%12)+1, not (m+1)%12 as it is
printed in the spec.