Made QueryTester db implementation independent.
Now the QueryTester works with any LDBC SNB DB implementation. Moved it
to a new module calls snb-interactive-tools.

The only drawback here is that to construct different implementations of
DbConnectionState generically necessitate either describing the
constructor parameter types and values in the configuration file, or
unifying the constructors to just take something like Map<String,
String>. I went with the latter approach.