Adding copy of QueryTester back to Neo4j.
Previously pulled the QueryTester out of Neo4j to make it a more general
tool for all the DB implementations. This unfortunately also makes using
the tool dependent on being able to compile all of the databases, which
most users won't be able to do because they lack the libraries necessary
to compile Torc (e.g. RAMCloud libraries).

So, I decided to add a copy of the QueryTester back into the Neo4j repo
so that users don't have to worry about this.