Changed Query10 to avoid Sack step.
Previously, Query10 was implemented using a series of Sack steps to
calculate similarity scores for friends-of-friends. Unfortunately this
prevented bulk fetching of vertices in TorcGraph via TorcVertexStep,
because the sack step assigns sack values one-by-one.

Reworked the query to avoid the sack step entirely, and allow for the
greatest degree of parallelism for TorcDB.