Changed file distr. in GraphLoader threads.
Previously each GraphLoader instance would generate a consistent list of all the
files for the social network and then partition it into contiguous segments and
assign segments to threads for loading. This was an OK strategy for the smaller
graphs, but for larger SNB datasets this resulted in some GraphLoader instances
getting a disproportionate amount of work to do. This was because files for
things like comment nodes are divided up automatically by the dataset generator
into many partitions, each of which is relatively large since there are so many
comments in the graph, and these files were listed contiguously in the file list
for the GraphLoaders. When taking contiguous segments of the file list to load,
some GraphLoader instances would get all the comment files to load while others
got much smaller files for loading.

One approach to solving this problem was to randomize the file list before
dividing it up among the workers, which could have been done consistently using
a deterministic seed.

The approach I decided to go with, however, was to make it so that for N
threads, each thread picks every Nth file to load in the list. So for 8 threads,
thread 3 in the system picks the 3rd file, the 11th file, the 19th file, and so
on.