Randomized file list in GraphLoader.
Randomized order of file list so that the load is evenly distributed among the
loader threads.

Note: A previous commit helped with this by interleaving the non-randomized file
list between all the threads. The problem with this approach is when loading
edges, where a long series of files in the list have to do with loading edges
for similar nodes. This means that when loading edges, we run the risk of
threads having their transactions conflict with each other. Randomizing the
order of files in the file list helps to solve this problem.