Made ImageMaker more efficient ico no prop edges.
Previously the ImageMaker would fill up a list with empty property maps, which
was taking up a lot of memory uselessly in the case of very large edge lists
that do not have properties. Changed this to not fill the list at all, and hand
an empty list to Graph.loadEdges.