Fixed bug in ImageMaker.
ImageMaker was using the SnbEntity names as neighbor labels, which is different
from what TorcDB uses, and also internally inconsistent in ImageMaker. Should
use the 'label' field of TorcEntity for vertex labels. This was being used in
the construction of the edge list key prefixes, but not in the neighbor label
lists.