Added default implementations for all complex queries.
After I implemented the "fakeComplexReads" feature, I registered all the complex
queries so that they would be able to run in fakeComplexRead mode. However, when
not in fakeComplexRead mode, the unimplemented complex reads would not return
anything to the driver, and the driver would get confused and eventually throw
an error.

Now all of the complex queries return something in non-fakeComplexReads mode, so
that the driver does not get confused. Note, however, that these queries will
run (and fail) during validation.