Fixed issue with repeat step in Query1.
It appears that the repeat step feeds traversers one by one through the loop, so
the barrier() step that I had in there wasn't batching up the vertices as I
expected and the out("knows") step was not batch fetching vertices. So, I
changed this to not use the repeat step.