Added locks to TorcDb2Server.
These locks make it so that only one query handler is being executed at one time. The reason for
this is that TorcDB2 is not thread-safe (one because the RAMCloud client is not thread-safe, and
two because two threads will share the same transaction context). So in TorcDb2Server we allow only
one query to execute at a time by wrapping the `executeOperation` method calls with a lock.

My longer term plan here is to make TorcDB2 support thread-local transaction contexts, and put
locks around the RAMCloud client method calls, probably in the RAMCloud C++ library itself.